New Development: No             JIRA Issue: CAS-3426         Ready for Test: Yes      Interface Changes: No What Interface Changed:          Test Programs: test_tflagger   Put in Release Notes: No              Module(s): Module Names change impacts.            Description: Catching when the output file to save the parameters 	   already exists.
git-svn-id: https://svn.cv.nrao.edu/svn/casa/active@17812 ac9c0da5-131f-0410-9009-a9c195f183e6

+++ code/xmlcasa/scripts/task_tflagger.py
@@ -426,11 +426,12 @@ def writeCMD(msfile, flagcmd, outfile):
 
     # Save to a text file
     if outfile != '':
-        try:
-            ffout = open(outfile, 'w')
-        except:
-            raise Exception, 'Error opening output file ' \
-                + outfile
+        if os.path.exists(outfile):
+            raise Exception, 'Output file already exists, ' \
+                            + outfile
+            
+        ffout = open(outfile, 'w')
+
         try:
             casalog.post('Will save the current parameters to '+outfile)
             print >> ffout, '%s' % flagcmd
